more css updates, platforms tweaks and componentize Measurements - pull Measurments section into it's own component - fix experiment platform selector - add white icons for platform selector in featured experiments - more css tweaks to featured experiment and update list
